fix(tracing): Honour disable metrics
Don't send traces to Sentry when users have opted-out of metrics. This
was previously only honoured by our in-house metrics solution.

For the CLI we have to attempt to parse the config earlier than normal,
defaulting to `false` if we can't, and only grabbing the single config
item. For the watchdog we pass over a new optional flag which is clearer
than depending directly on the environment variable.

I thought about flipping the logic to be `enable` rather than `disable`
but it was harder to reason about in the context of disabling metrics
outside the fn.