fix: uninitialized memory in pkgdb arg parser
This argument was never initialized, and its value was only ever set to
`true` when building a container e.g. it was never explicitly set to
false. When the uninitialized value is 0, everything accidentally works
as intended. When the uninitialized value is non-zero, then `buildenv`
interprets this as having been passed the `-c` argument and builds a
container for the lockfile. This breaks the environment because the
contents of the store path do not match those of a rendered environment.