feat(buildenv): include manifest lock in environment
Write the lockfile to the environment,
causing semantic changes to the lockfile to invalidate existing environment store paths.

This is useful to determine whether an environment needs to be rebuilt,
as well as reresolve the environment "linktree".
The latter is particularly important when removing package priorities,
that previously resolved package conflicts.

Since the `priority` is not part of the evaluation,
setting the `priority` for a package does not change the evaluated store path
of the environment.
Removing the `priority` will built the previous environment
and _not_ error on file conflicts,
until another change to the manifest causes a rebuild,
or the environment is built on another machine.