fix: socket conflict for layered activations
We currently set `_FLOX_SERVICES_SOCKET` when we activate an environment
with services. We also check `_FLOX_SERVICES_SOCKET` when running an
activation to allow overriding the path for the socket. This means that
when an activation is performed within another activation, it uses the
value of `_FLOX_SERVICES_SOCKET` set by the outer activation, which
causes services for two different environments to interfere.

Instead, introduce a second variable `_FLOX_SERVICES_SOCKET_OVERRIDE` to
provide the overriding behavior, and use `_FLOX_SERVICES_SOCKET` to
communicate with the activate script.

We could instead drop the one test that uses the overriding behavior,
but I figure it could be useful in debugging to be able to set the
socket path explicitly.