fix(tracing): Enable tracing for activate
Make this work by explicitly:

- ensuring there's a single span, because we can't find parents
- closing the active span, otherwise nothing will be flushed
- closing the client, to flush the closed span

This makes it easier to add new tags and potentially allows us to
stop maintaining two metric implementations in the future.

The main difference between our previous attempt is that we're not
instrumenting the parent functions, which causes spans that we can't
close and prevents the transaction from being flushed. `span.exit()` is
just a more explicit form of `drop(span)`.

The timeout is set low enough that we forfeit data instead of degrading
the experience of `flox activate` too much if we're unable to flush.

In my local testing there appears to be no impact to the execution time
if you don't have _any_ internet connection. It's possible that slow
connections could cause a delay before handing off to the command but I
don't believe this would be any different to the automatic flush on Drop
for other commands.