test(pkgdb): use external flake to avoid failure when repo is in dirty
`test_locksUrl` failed during local development due to using a local flakeref
which references the local (containing) git repository through `localTestFlake`.

`localTestFlake` is defined as `./tests/data/lock-flake-installable`,
which is a subdirectory within the flox repo.

If there are any tracked changes, the repo is considered dirty.
This propagates to any flake defined within the repo.
Thus in development this test would fail most of the time,
unless the repo is in a clean state.
To avoid this, exceptionally use an external flake for this test,
that can be locked cleanly.