feat: builder implementation
Define a `ManifestBuilder` trait that provides an interface to building packages.
The signature of this trait is inspired by the requirements
of calling the `flox-build.mk` subsystem, outlined in
<https://github.com/flox/flox/issues/1998>.
Namely, builds require
* a `basedir`: the directory containing the .flox of the package for reference
* a rendered `flox_env`: a symlink or realpath to the built environment
  which contains build instructions and build metadata.
* the names of packages to build

The build will run in the background but can be observed using the returned `BuildOutput`.
`BuildOutput` implements an iterator over events from the build,
currently lines of `std{out,err}` and an `Exit` event to wait for termination or the build.