chore: remove watchdog helper script
Polling for the watchdog had to be added in a few places, so
wait_for_watchdog was factored out.

It was particularly confusing why tests were flaking before adding a
`wait_for_watchdogs` before `project_teardown`. Our best guess is that
process-compose down is hanging forever when .flox, which contains its
logfiles, is removed.