feat(watchdog): Log a heartbeat every hour
So that the upcoming garbage collection of log files can determine
whether a watchdog log file is still in use without needing to lookup
the respective watchdog process.

The alternative to this was to extend the environment registry to track
the watchdog PID in addition to the activation PID, because we can't
lookup the former from the latter once the activation has exited and the
watchdog has been re-parented, but it would be more work than this and
we don't currently have another need for it.

This is spawned in the background before we block on waiting for
termination and dies with the watchdog on shutdown. 1hr seems like a
reasonable balance between generating too many and too little log
entries.