feat(watchdog): GC watchdog and services logs
To prevent the recently added `.flox/log` directory from growing too
large whilst still keeping enough logs that users can use to inspect and
report issues. We will now keep:

- Watchdog logs that have been updated within the last 3 days.

A new file is created every time a watchdog is started, which happens
with every non-ephemeral activation. The file haves a PID in their name
but we don't parse it.

This relies on the preceding commit that emits a heartbeat log entry
every hour so that we can determine whether a watchdog is still using
the file without needing to look it up or check open files.

A side-effect of this is that upgrading from 1.3.0 will blow away the
watchdog logs for any activations that were started more than 3 days ago
because they won't have the heartbeats. The watchdogs will still be able
to continue writing to the unlinked files and shutdown correctly though.

- Services logs from the last 5 invocations.

A new file is created every time `process-compose` is started with
`activate --start-services`, `flox services start`, or `flox services
restart`. The files have a timestamp in their name so they can be sorted
by oldest to newest.

I briefly explored using `lsof` to exclude currently open files, so that
we wouldn't need to maintain separate heuristics for each type of log
file that we have now and in future, but the open-source binary was
dramatically slower on MacOS (~3s) and it felt cumbersome to integrate
with. So we'll have to add explicit support for other types of log files
that we introduce in the future.

All of the outer functions err on the side of reporting and continuing
on errors so that we don't get stuck on a "poison pill" file that
prevents the deletion of other files. This could happen for instance if
someone runs an activation and watchdog under a different user but it
would correct itself if there were subsequent runs under the same user.
I'm not able to simulate this in a unit test though.

The deletion is started in a background thread fairly early in the main
thread, once we've verified that we can run and before we block on the
signal watchers. The thread will then naturally exit, regardless of how
far it got, when the main thread shuts down.

Concurrent deletions from multiple watchdogs should also be fine whilst
they use the same criteria; we'll need to bear it in mind when changing
the criteria though.