feat: add libsandbox.{so,dylib} to package-builder
The sandbox library will required by the manifest builder in the second
phase of the manifest build project, and while it's not required in the
first phase it's less work to simply import the PoC unchanged and then
we can elect to use it (or not) from within the manifest build makefile.

Note that the sandbox library may be of use for ld-floxlib.so as well
in future, but for not it's only required from the package builder so
we instead maintain it in the package-builder directory.