fix(activate): Warn for -s if there are none
Print a warning if you `flox activate --start-services` on an
environment that doesn't have any defined.

Previously we would continue and silently not start services, which
caught me out when manually testing something on an environment that I
thought would start services but didn't because I used the wrong
manifest.

Currently opting for warning and continuing instead of an error and
exiting because you might be iterating manifest changes and re-running
the same command knowing that it won't fully work.