chore: (also a fix) build flox with only one nixpkgs revision
... and let that revision be derived from a) the flox/nixpkgs clone
and b) from its "stable" branch that automatically updates over time,
and c) using the nixVersions.stable version of nix.

This of course will require that we take the proactive approach of
building flox against the staging and unstable branches ahead of time
so that we can be aware of problems as they arise from upstream, but
that's the subject for a different discussion.

Note that we're temporarily pinning at staging.20240817 while we sort
out a better approach for process-compose versioning.