fix: re-render ACTIVATION_SCRIPTS_PACKAGE_DIR with flake
The pkgdb Makefile was not picking up changes to the activation
scripts package when inputs were changed, which of course can
cause changes in the string-substituted versions of activation
scripts as created by the Nix package build. This change updates
the conditions for re-rendering the package to include changes in
either/both of the top-level flake.{nix,lock} files.