chore: build flox using github:flox/nixpkgs/stable only
Update nixpkgs input to be derived from a) the flox/nixpkgs clone
and b) from its "stable" branch that automatically updates over time.

This of course requires that we proactively build flox against the
github:flox/nixpkgs staging and unstable branches ahead of time so
that we can be aware of problems as they arrive from upstream, well
before they hit the stable branch.