feat: err/warn for services that are not enabled on the current system
* Add `ManifestServices::copy_for_system` to create a new ManifestServices instance
with services for systems other than provided `system` filtered out.

Clone the services rather than filter in place
to avoid accidental mutation of the original in memory manifest/lockfile.

`copy_for_system` was implemented on the manifest type
rather than in the `services` module to avoid leaking structure
and construction of `ManifestServices`.

* Where previously checking `manifest.services.is_empty()`
add an additional check whether `copy_for_system` is empty as well and warn/err
if the manifest provides no service for the current system.
The messages will be different from _not having services at all_ for clarity.

* Check service names against services for current system
When providing services as an argument, the service name
will now be checked against `copy_for_system` and raise an error
if the service is not available.

* Only resolve to supported services if all services are implied.
`flox services *` will substitute only supported services
for the current system if not explicit service names are provided.