chore(CI): move build of flox-cli-tests to nix-build step
The `remote` job currently builds both `flox` and `flox-cli-tests`,
which adds ~45 sec to every `remote` job. Most of this work is wasted,
since it copies flox to the GitHub runner, even though flox is only
needed on our own runner for that job.

Building `flox-cli-tests` in the `nix-build` job is much faster, because
that job has already built `flox`. Building `flox-cli-tests` is
relatively cheap since it only requires a `writeShellScriptBin`.

The only jobs that depend on `nix-build` are `remote`,
`Build installers`, and `Report Falure`. We skip `Build installers` on
PRs and in the merge queue, so adding a smaller amount of time to
`nix-build` in exchange for subtracting a greater amount of time from
`remote` should be an improvement.