fix: support fish 3.2.2
fish 3.2.2 fails due to our usage of return outside of a function and
`$(...)`

Before this commit, the test added in this commit fails with:
```
   /private/var/folders/2w/r9qm37r12qs4xn3sjn4cv2ym0000gn/T/nix-shell.JcSQ0t/bats-run-MtPfNW/test/1/project-1/.flox/run/aarch64-darwin.project-1/activate.d/fish (line 17): 'return' outside of function definition
     return 1
     ^
   from sourcing file /private/var/folders/2w/r9qm37r12qs4xn3sjn4cv2ym0000gn/T/nix-shell.JcSQ0t/bats-run-MtPfNW/test/1/project-1/.flox/run/aarch64-darwin.project-1/activate.d/fish
   source: Error while reading file “/private/var/folders/2w/r9qm37r12qs4xn3sjn4cv2ym0000gn/T/nix-shell.JcSQ0t/bats-run-MtPfNW/test/1/project-1/.flox/run/aarch64-darwin.project-1/activate.d/fish”
```

Fix this by using exit 1 instead of return 1.

After changing that, there is still an error for:

`$(...) is not supported. In fish, please use '(_gnused/bin/sed)'`

Fix this by piping to source.