feat: render top-level requisites.txt file to flox environments
Create `requisites.txt` containing a list of all packages in the closure
as a file in the top directory of the rendered flox environment.

This file can be useful for a variety of reasons, but we are adding
it at this stage in preparation for the upcoming flox manifest build
"sandbox" features which requires a way to discern the Nix packages
that are in and out of the environment closure without having to
first query narinfo data from the Nix database.

Co-authored-by: Matthew Kenigsberg <matthew@floxdev.com>