chore: race condition in test
chore: poll for service states in unit test

The current timeout was 10ms which may have been a tad too little
to avoid race conditions in all cases.
My understanding is that we are moving to polling in tests,
rather than just adding longer timeouts, so that's what we do here.