refactor(handle_pull_result): return deferred message context
Defer printing of the final message during result handling,
to allow conditionally displaying different headers in preparation
of adding support for `--copy`.

Copying (as implemented in the following commit) entails
1. pulling a `ManagedEnvironment`
2. handling the result of pulling the environment
   (e.g. migrating v0 manifests, adding the current system as a supported system)
3. converting the environment to a path environment

When pulling an environment regularly, the header of the message is expected to read:

    Pulled owner/name from hub.flox.dev

While pulling with `--copy` according to <https://github.com/flox/flox/issues/1226>,
is expected to terminate with:

    Created path environment from owner/name

Assuming immediate printing in 2. while respecting the current mode,
would result in the right message being printed _before_ the conversion (3.).
Consequently, the messaging would not be consistent with reality,
if the conversion was to fail.

Alternatively, converting the environment _before_ handling errors,
fails on account of migration functionality currently only being implemented
for `ManagedEnvironment` and requires catching addressable errors twice;
Once before converting to a `PathEnvironment` and a second time
during migration/incompatibility handling.

The approach chosen here, defers the printing of the final message
until after the (conditional) conversion step,
where the final header is added to the context deferred from handling the pull result.

Preferably our messaging would be more linear, allowing for context to be printed
in the order it appears during execution, rather than being accumulated and deferred
into a final custom message.