test: add set -x for overmind test
The `kills daemon process` test flaked in
https://github.com/flox/flox/actions/runs/10742333592/job/29794735526#step:8:297

It's unclear why that run flaked. `overmind status` printed something,
and it doesn't seem to have ever failed since there is no output for
failure to connect.

It's odd that PID 0 was printed, but when run locally, that happens
occasionally when `overmind start` and `overmind status` are racing.
Locally, that still exits 0.

So it seems `overmind status` should have succeeded, but then timeout
still timed out.

Add set -x to help with debugging in the future.