fix: deserialize MessageType's other than Unknown
456cc7b07052e84571d514b395609f608a962850 made MessageType an untagged
enum so that the Unknown variant would deserialize correctly, but this
meant that all strings would deserialize as Unknown rather than being
deserialized into the correct variant. "constraints_too_tight" for
example was being deserialized as `Unknown("constraints_too_tight")`.

The Unknown variant should be untagged, but the enum as a whole should
not be.