fix: remove leading whitespace in `ps` output
On linux the output of `ps` starts with two space characters.
Consequently, the `cut -d' ' -f1` used to extract the pid column,
will extract an empty field.

This causes the loop polling for the watchdog to break on the first
iteration rather than actually waiting for the watchdog.

wait_for_watchdogs was factored out for use in the following commit