refactor: add `systems` field to `PackageToInstall::CatalogPackage`
Allow installing packages for specific systems only.
Currently installing packages through `Environment::install`
always installs packages for _all_ systems.
In case a package is not available on all platforms,
the package can only be installed by manually editing the manifest.

This commit adds a `system` field to the `PackageToInstall` variant
for catalog packages, to allow callers to amend the systems
that the package is installed for.
Since we do not have a formal syntax to expose this on the CLI,
the `FromStr` implementation remains unchanged and will not explicitly
restrict the systems in any way.