feat: attempt install with reduced systems
During `flox install`, when all resolution errors are "package unavailable on some systems",
modify the `systems` attribute in the original list of packages to install,
to match the suggested systems returned by the catalog service
and attempt to install again.

* Warn if the catalog service returned nonsensical data.
That is resolution errors for install-ids that were not requested,
or resolution errors for non-catalog packages.

* test: amend tests for unsupported packages

* test: cannot install with reduced systems with other failures
When a resolution fails with errors other than missing systems,
do not attempt installing with suggested constraints.
Instead, format the error as before.

* test: amend tests for unsupported packages

* test: cannot install with reduced systems with other failures
When a resolution fails with errors other than missing systems,
do not attempt installing with suggested constraints.
Instead, format the error as before.