feat(pkgdb): add `--container-tag` option
Add the `--container-tag` option to `pkgdb buildenv`.
This allows buildenv to generate a container image with a specific tag.

E.g.

```console
$ pkgdb buildenv "$some_lock_file" \
  --container 'foo' \
  --container-tag 'bar'
```

Will produce a image that, when loaded with a tool like Docker, will be called `foo:bar`.

If the flag isn't passed along, a default of `latest` will be used.