refactor(watchdog): Replace prctl with procfs
So that we can later support running a single watchdog for all
activations of an environment and monitor all of their PIDs at once.

`prctl` for Linux can only monitor the parent PID. `kqueue` for MacOS
can already monitor multiple PIDs and there's no `procfs` available so
we don't need to make any changes there.

Using `procfs` rather than `kill -0` so that we can differentiate zombie
processes that haven't been reaped by their parents yet.

Implemented reading `procfs` directly rather than pulling in the
`procfs` crate because the part that we is very simple, it's already
awkward enough to make changes on MacOS without have platform specific
dependencies, and I wanted some tests that we were filtering the correct
states anyway.

One thing that the `procfs` crate does differently is that it will hold
an open file handle on instantiation for a PID in case it's replaced by
another process. It's not a deal breaker for us though because we should
notice within the polling interval and we're going to have to reload the
list of PIDs as activations come and go.

There's no explicit support for watching multiple PIDs yet but it should
now be simple to add for both implementations and we have a separate
ticket (2129) for it in the current sprint.

The SIGUSR1 handler was serving two purposes; both for `prctl` and for
manual debugging/testing, as exercised by a separate integration test,
so that has been left as-is.