test(activate): Fix setup_user_dotfiles guard
This wasn't working because it was missing line continuations. It should
have failed when running `user_dotfiles_setup` for every test in the
preceding commit but it didn't.

I have absolutely no idea what `bash` is actually interpreting it as and
why it doesn't cause either a syntax error or non-zero return.

Replace it with the `noclobber` option which will cause a non-zero
return if any of the files that are being redirected to already exist.
This also means that we don't have to maintain two lists of files.

The new error that this generates will be fixed in a following commit:

    ✗ zdotdir: test zshrc activation [5]
       tags: activate activate activate:zdotdir activate:zdotdir:zshrc
       (from function `user_dotfiles_setup' in file activate.bats, line 32,
        from function `setup' in test file activate.bats, line 121)
         `user_dotfiles_setup' failed
       /var/folders/8q/spckhr654cv4xrcv0fxsrlvc0000gn/T/nix-shell.DPCf5O/flox-cli-tests-CnGz2O/activate.bats: line 32: /var/folders/8q/spckhr654cv4xrcv0fxsrlvc0000gn/T/nix-shell.DPCf5O/bats-run-TCOCWA/suite/home/.profile: cannot overwrite existing file
       /var/folders/8q/spckhr654cv4xrcv0fxsrlvc0000gn/T/nix-shell.DPCf5O/flox-cli-tests-CnGz2O/activate.bats: line 106: popd: directory stack empty