test(activate): Isolate $HOME between tests
This is the last step to ensure that we don't leak state between tests
that make changes to `$HOME`. Now that we have a new `$HOME` for each
test we no longer need to explicitly cleanup the `.extra` files.

This results in a small increase to the total runtime but we should be
able to claw that back with parallelisation now that we are running the
tests in a safer way.

Before on `main`:

    106 tests, 0 failures in 49 seconds

After:

    106 tests, 0 failures in 62 seconds