test(activate): Remove NO_PARALLELIZE_WITHIN_FILE
This should be safe to enable now that we're isolating `$HOME` between
tests. It's a partial revert from when we enabled parallelisation in
7bc3d0c3.

This gives us back the time that we lost by isolating `$HOME` for each
test plus a little bit more.

Before on `main`:

    106 tests, 0 failures in 49 seconds

Before with isolation:

    106 tests, 0 failures in 62 seconds

After:

    106 tests, 0 failures in 35 seconds