test(activate): Test zsh in-place from RC files
Test that in-place activation works from zsh RC files. This is a
variation of the tests above it which do something similar but not quite
the same by hooking into the RC process later on.

I was hoping to reproduce the issue in 1946 but I haven't been able to
either manually or with these tests. If the problem still exists from
Graham then we may be able to tweak these new tests to reproduce it.