feat: Add flox-activations crate
This also adds a new build profile called "small" that is tuned to
produce smaller binaries. Using this profile reduces the size of the
compiled binary from 995k ("release" profile) to 471k. This application
shouldn't be doing anything performance intensive, so the difference in
speed should be negligible in practice.