fix: zsh in-place activation in dotfile
A `flox activate` of one environment behaves incorrectly if .zshrc
contains `eval "$(flox activate)"` of another environment.

The outer `flox activate` passes `FLOX_ENV` to our activate.d/zsh script
as an environment variable. But before activate.d/zsh is run, zsh will run
activate.d/zsh/.zshrc, which sources the user's .zshrc. If the user's
.zshrc contains an in-place activation, we export FLOX_ENV. Then when
activate.d/zsh is run for the outer activation, it has the incorrect
value of FLOX_ENV set.

To avoid this, save the value of `FLOX_ENV` in activate.d/zdotdir/.zshrc
and restore it after sourcing the user's .zshrc.

Do the same for all other variables we export for an in-place
activation. I haven't checked if this is strictly necessary for all of
them, but it shouldn't hurt.