refactor: move Versions to a new shared crate
We'll want to share code between `flox-activations` and
`flox`/`flox-rust-sdk`. `flox-activations` should not depend on
`flox-rust-sdk` because it needs to be of minimal size. `flox-rust-sdk`
could depend on `flox-activations`, but it seems that would be more
disorganized, and it doesn't seem that it will be additional effort to
move helpers from `flox-rust-sdk` to `shared` as opposed to moving them
from `flox-rust-sdk` to `flox-activations`.