test: add message `History` to track messages sent by the CLI in tests
A history for messages printed to stderr through the `message` module .
In unit tests, the messaging functions of the `message` module will,
populate a `history` in addition to printing the message to stderr.
this allows the test to assert against the messages printed to stderr,
without refactoring existing implementations to print to a `Write` trait
and injecting a mock in tests.

The `History` is thus comparatively non-intrusive.
However, it comes with a few caveats:

Messaging functions are currently effectively (process) globals,
inheriting the global nature of `stderr`.
`#[test]` functions are run in parallel on separate _threads_
by both `cargo test` and `cargo nextest` to some degree.
Assuming an equally global `History`, messages from one test
will be visible in another test.
There are several approaches to avoid this:

* running tests in sequence (e.g. `cargo test -- --test-threads=1`),
  slows down the test suite.
* additional bookkeeping of scopes (e.g. via `tracing`),
  complicates the test setup, wither needs to be done manually,
  or we maintain a `#[test]` replacement. (That said,
  there may be something to learn from `tracing`s architecture).
* Pass a `History` instance to the messaging functions,
  intrusive as it reqeuires all callers to be updated.
* Use a [thread_local!] `History`, which is the current approach.

Being thread local, the `History` is not shared between threads,
i.e. not shared between tests.
On the flip side, if the test spawns additional threads,
which print messages, these will be captured in the `History`
of the thread that spawned them, rather than the test's `History`.
Likewise if the test uses a multi-threaded async runtime, like `tokio`,
messages from spawned tasks will not be captured in the `History`.
Also due to runtimes sharing threads between tasks, the `History` available
in the task is also shared between tasks.

Luckily, our CLI surface is largely single threaded.
Threads are mainly spawned for blocking operations either in the library layer,
or to run library functions in the background, e.g. to print a progress bar
on the main thread.
Also, while technically within an async runtime, the CLI is largely synchronous,
and the default [#[tokio::test]] runtime is single threaded.

* refactor: Use an `Rc<RefCell<_>>` as the data backing.
Allow `History` instances to
1) allow for multiple histories with the same API (e.g. a separate history for stdout messages)
2) provide a handle for convenience methods like "contains *" checks.