feat: support attributes with leading `@`
Previously the first `@` character found in an install argument was interpreted
as the delimiter of attr_path and version.

However, nixpkgs _does_ contain several packages with attributes starting with `@`.
For example, `nodePackages.@angular`.
Previously `@angular` was incorrectly interpreted as the version `angular`.
This commit adds support for such leading `@` characters in attributes.

In particular we will now search for the first `@` character,
but instead of eagerly splitting and returning a version,
if that `@` is the first character of an attribute, (the first character,
or immediately following a `.`), we continue searching for further `@` characters.

* refactor: replace recursion with loop
* feat: support version delimiters in descriptors that start with `@`.