fix(build): Always re-render build script
The previous fix to retain the previous build script had the effect of
requiring that it always be older than the corresponding build script
source file in order to be regenerated, but that doesn't work so well
when the source file is coming from the Nix store with a timestamp of 1.

Adding the `FORCE` prerequisite will force the rebuilding of the target
in each instance, and we correspondingly ignore that value when iterating
through the prerequisites within the shell loop.

Also adding logic to write new version to a separate file so that we can
atomically move it into place.