test(build): Preserve arg0 for wrapped binaries
So that usage strings which depend on arg0 display:

    Usage: xh <[METHOD] URL> [REQUEST_ITEM]...

Rather than:

    Usage: .xh-wrapped <[METHOD] URL> [REQUEST_ITEM]...

This isn't possible for interpreted scripts or variations of
`/proc/self/exe` but I've documented the those behaviours in tests for
the avoidance of doubt.

For the same reason, it's not possible to chain `exec -a` into
`activate` because it's an interpreted script and `$0` always evaluates
to `activate`.

However we already export `FLOX_SET_ARG0` in our top-most wrapper
script. `activate --turbo` now switches on this being present, whilst
still preserving the existing behaviour of being able to call shell
builtin commands.

I've used Go rather than Rust for the example projects because we
already had catalog mocks and it matched the original reproducer from
`charmbracelet/vhs`.

It's borderline whether these tests should be "integration tests" rather
than "unit tests" because they're doing a complete round-trip and
depending on the `flox` binary. So any time you make a change to the
activation script you also have to rebuild the CLI in order for the
tests to pick it up.