feat(watchdog): wait for all pids on macos
This adds a trait for a generic "watcher" that waits for a termination
condition. This commit implements it for a watcher on macOS that
terminates once all of the PIDs that it's watching have terminated.

Initialization of kqueue doesn't happen until the
`Watcher::wait_for_termination` method is called, which requires the
inner kqueue watcher to be stored as an `Option` that's later checked at
runtime. I think it's impossible for this to ever end up as `None` after
calling `wait_for_termination`, but a follow up PR could make
initialization happen in two stages so that the `Option` isn't
necessary.