fix(watchdog): prevent startup failure
This adds a check on startup that could prevent a watchog from starting
if the previous watchdog failed before cleaning up PIDs in the registry.
This check is (hopefully) reliable on Linux, but best-effort on macOS
due to the available OS APIs. See the comments in
`should_bail_on_startup` for details.