feat: new pkgdb realisepkgs subcommand
This commit adds a new `pkgdb realisepkgs` subcommand for realising pkgs
referenced from within a manifest lockfile. It is used in conjunction
with the nix-based `buildenv` as detailed in the following commit.

The tests added with this commit are copied largely from the corresponding
tests for `pkgdb buildenv`, with the expectation that we will be removing
all code and tests associated with this pkgdb subcommand in future.