feat: impl `Activation` modifiers
Add methods to retrieve the activations id and readiness state,
as well as methods to add or remove registered PIDs.

Non of the attributes of an `Activation` are meant to be modified directly.
`ready` is modified but only in a "one way" manner, i.e. once ready
activations do not return into a not-ready state.
`attached_pids` is modified, but mutable access to the storing `Vec`,
could break assumptions (e.g. of order) and exposes another internal struct unnecessarily.