feat: read and write activations.json with file lock
refactor: acquire file lock to read activations.json
Return a file lock that can be used to write the modified activations.json file.
Locking in the same step as reading is supposed to reduce TOCTOU kind of issues,
where the reading lock is dropped and anew write lock is acquired later.