test(activate): Run shfmt on build
To ensure that we don't regress formatting after applying it in a
preceding commit. This is limited to activation scripts for now, like
`shellcheck` is, because we haven't dealt with formatting of other
scripts in the repo.

This misses a small number of files that don't have an extension or
shebang. I did experiment with renaming the `{bash,zsh,tcsh,fish}` files
to have extensions but `shfmt` incorrectly attempts to format `.zsh`
files at the moment (PR incoming). Also naming them is hard. So we'll
live with this for now.

We temporarily need `.editorconfig` so that `shfmt` can apply the same
rules that our editors will use. `shfmt` discovers this file
automatically and doesn't have any options for specify its path.

The `chmod` is necessary to prevent the following error because the
directory is copied with `0555` by default:

    cp: cannot create regular file '/nix/store/gvgp6a7m21ifmc55794frnff16y19nzq-flox-activation-scripts/.editorconfig': Permission denied