refactor(watchdog): unify macos/linux impls
This removes the kqueue watcher from macOS in favor of just calling
`/bin/ps` since attempting to watch a terminated process will cause an
error. The macOS and Linux implementations turned out to be basically
identical, so they were unified into a single struct.

I considered using `libproc.h` on macOS, but it's technically a private
API and the build was tricky to set up. I'd like to use that eventually
instead of parsing the output of `ps`, but for expediency I decided
against it for now.