test(build): Go mocks for non-Darwin
These tests would fail on Linux systems because they were using a
multi-part mock response based on `flox init` where the initial request
only resolves for Darwin. This results in the environment that silently
omits the package and no `$FLOX_ENV/bin` is generated:

    stdout: make: Entering directory '/tmp/.tmplf5xse/temp/.tmpTH6NaM'
    stdout: Rendering foo build script to /tmp/nix-shell.5Qgf1P/def85939-foo-build.bash
    stdout: Building foo-0.0.0 in local mode
    stdout: MAKEFLAGS= out=/tmp/store_def85939439f94b0d8842804b8c5f21a-foo-0.0.0 /tmp/.tmplf5xse/temp/.tmpTH6NaM/.flox/run/aarch64-linux.name/activate --turbo -- bash -e /tmp/nix-shell.5Qgf1P/def85939-foo-build.bash
    stderr: /tmp/nix-shell.5Qgf1P/def85939-foo-build.bash: line 4: go: command not found
    stderr: make: *** [/mnt/package-builder/flox-build.mk:363: foo_local_build] Error 127
    stdout: make: Leaving directory '/tmp/.tmplf5xse/temp/.tmpTH6NaM'
    test providers::build::tests::build_wraps_binaries_with_preserved_arg0 ... FAILED

I ran into it this running the tests in a Linux container. Bill
originally reported this and I _believe_ he's on Linux:

- https://flox-dev.slack.com/archives/C05P6A5J6U8/p1728415411448669?thread_ts=1728415411.448669

What I'm not clear about is why the `cli-dev` job on the Ubuntu Runner
in CI hasn't been failing in the same way. This will be made more
reliable by:

- https://github.com/flox/flox/issues/2228
- https://github.com/flox/flox/issues/1659